# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APZ Legal Practice Management System API
security:
  - sessionCookie: []
  - bearerSession: []
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [individual, corporate, trust, government]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, pending]
        - name: ficaStatus
          in: query
          schema:
            type: string
            enum: [compliant, pending, expired, blocked]
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"

    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Archive a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Archived
  /clients/{id}/fica:
    get:
      operationId: getClientFica
      tags: [clients, fica]
      summary: Get FICA compliance status for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: FICA status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FicaStatus"
  /clients/{id}/risk-score:
    get:
      operationId: getClientRiskScore
      tags: [clients]
      summary: Get risk score for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"

  # ─── MATTERS ─────────────────────────────────────────────────────────────────
...
  /clients/{id}/compliance:
    patch:
      operationId: updateClientCompliance
      tags: [clients]
      summary: Update compliance fields (compliance roles only, audited)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientComplianceInput"
      responses:
        "200":
          description: Updated client.
        "400":
          description: Invalid fields.
        "403":
          description: Compliance role required.
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
    bearerSession:
      type: http
      scheme: bearer
      description: Opaque staff session token, not a JWT.
  schemas:
    ClientComplianceInput:
      type: object
      additionalProperties: false
      properties:
        complianceStatus:
          type: string
          enum: [compliant, review_required, blocked]
        complianceBlockReason:
          type: ["string", "null"]
          maxLength: 2000
        riskLevel:
          type: string
          enum: [low, medium, high]
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
...
    Client:
      type: object
      required: [id, name, type, status, ficaStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [individual, corporate, trust, government]
        status:
          type: string
          enum: [active, inactive, pending]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        idNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        riskScore:
          type: ["integer", "null"]
        ficaStatus:
          type: string
          enum: [compliant, pending, expired, blocked]
        companyRegistration:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        matterCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [individual, corporate, trust, government]
        email:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        address:
          type: string
        companyRegistration:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
    FicaStatus:
      type: object
      required: [clientId, overallStatus, documents]
      properties:
        clientId:
          type: integer
        overallStatus:
          type: string
          enum: [compliant, pending, expired, blocked]
        riskScore:
          type: integer
        documents:
          type: array
          items:
            $ref: "#/components/schemas/FicaDocument"
        nextReviewDate:
          type: ["string", "null"]
    FicaDocument:
      type: object
      required: [id, type, status]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [id_document, passport, proof_of_address, company_registration, beneficial_ownership]
        status:
          type: string
          enum: [verified, pending, expired, missing]
        expiryDate:
          type: ["string", "null"]
        uploadedAt:
          type: ["string", "null"]
    RiskScore:
      type: object
      required: [clientId, score, level, factors]
      properties:
        clientId:
          type: integer
        score:
          type: integer
        level:
          type: string
          enum: [low, medium, high]
        factors:
          type: array
          items:
            type: string

    # ─── MATTER ───────────────────────────────────────────────────────────────
...